Rename `apps` to `Applications`
