+++ kernel/Applications/animation.hpp
