+++ kernel/Applications/spinner.hpp
